updated algo to fit within 1 hr time window, and few other minor optimizations

+++ slack_helper.py
@@ -8,26 +8,37 @@ from slack_sdk.errors import SlackApiError
 from campaign import CampaignBuilder
 
 file_path = 'campaigns.csv'
+channel_id = 'C07MPL2QFKR'
 
 def send_message(campaigns, campaign_notes, st_time, end_time):
     slack_token = os.getenv('SLACK_TOKEN')
     client = WebClient(token=slack_token)
 
-    write_campaigns_to_csv(campaigns, campaign_notes)
-    try:
-        response = client.files_upload_v2(
-            channel='C07MPL2QFKR',
-            file=file_path,
-            title="📅 Upcoming Campaign Schedule",
-            initial_comment=f"🌟 Here are the campaign details which were scheduled between {st_time.strftime('%d %b %Y %I:%M%p')} and {end_time.strftime('%d %b %Y %I:%M%p')}, along with our suggested timings"
-        )
-        print(response)
-    except SlackApiError as e:
-        print(f"Error sending slack message: {e.response}")
-    finally:
-        if os.path.exists(file_path):
-            os.remove(file_path)
-            print(f"CSV file {file_path} deleted.")
+    if len(campaigns) == 0:
+        try:
+            response = client.chat_postMessage(
+                channel=channel_id,
+                text=f"ℹ️ No campaigns were found between {st_time.strftime('%d %b %Y %I:%M%p')} and {end_time.strftime('%d %b %Y %I:%M%p')}"
+            )
+            print(response)
+        except SlackApiError as e:
+            print(f"Error sending slack message: {e.response}")
+    else:
+        write_campaigns_to_csv(campaigns, campaign_notes)
+        try:
+            response = client.files_upload_v2(
+                channel=channel_id,
+                file=file_path,
+                title="📅 Upcoming Campaign Schedule",
+                initial_comment=f"🌟 Here are the campaign details which were scheduled between {st_time.strftime('%d %b %Y %I:%M%p')} and {end_time.strftime('%d %b %Y %I:%M%p')}, along with our suggested timings"
+            )
+            print(response)
+        except SlackApiError as e:
+            print(f"Error sending slack message: {e.response}")
+        finally:
+            if os.path.exists(file_path):
+                os.remove(file_path)
+                print(f"CSV file {file_path} deleted.")
 
 
 def write_campaigns_to_csv(campaigns, campaign_notes):
